Fix compiler warnings
git-svn-id: https://svn.code.sf.net/p/prpnet/code@87 029e2237-403a-42e8-9450-7768c132b01b

+++ source/Socket.cpp
@@ -53,7 +53,7 @@ SharedMemoryItem  *Socket::GetThreadWaiter(void)
 {
    char   id[10];
 
-   sprintf(id, "TW %d", ii_SocketID);
+   sprintf(id, "TW %d", (int32_t) ii_SocketID);
 
    if (!ip_ThreadWaiter)
       ip_ThreadWaiter = new SharedMemoryItem(id);
